@@ -8,24 +8,28 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
-        "fideloper/proxy": "4.4.1",
-        "fruitcake/laravel-cors": "2.0.3",
-        "guzzlehttp/guzzle": "7.2.0",
-        "laravel/framework": "7.30.1",
-        "laravel/tinker": "2.5.0",
-        "laravel/ui": "^2.1",
-        "lorisleiva/laravel-deployer": "0.3.5",
-        "sentry/sentry-laravel": "^2.0"
+        "php": "^7.3|^8.0",
+        "doctrine/dbal": "^3.0",
+        "fideloper/proxy": "^4",
+        "fruitcake/laravel-cors": "^2",
+        "guzzlehttp/guzzle": "^7",
+        "laravel/framework": "^8",
+        "laravel/sanctum": "^2",
+        "laravel/tinker": "^2",
+        "laravel/ui": "^3",
+        "lorisleiva/laravel-deployer": "^0",
+        "sentry/sentry-laravel": "^2",
+        "spatie/laravel-activitylog": "^3.16"
     },
     "require-dev": {
-        "facade/ignition": "2.5.8",
-        "fzaninotto/faker": "1.9.2",
-        "mockery/mockery": "1.4.2",
-        "nunomaduro/collision": "5.1.0",
-        "pestphp/pest": "^1.0",
-        "pestphp/pest-plugin-laravel": "^1.0",
-        "phpunit/phpunit": "9.5.0"
+        "facade/ignition": "^2.5",
+        "fakerphp/faker": "^1.9.1",
+        "laravel/sail": "^1",
+        "mockery/mockery": "^1.4.2",
+        "nunomaduro/collision": "^5.0",
+        "pestphp/pest": "^1",
+        "pestphp/pest-plugin-laravel": "^1",
+        "phpunit/phpunit": "^9"
     },
     "config": {
         "optimize-autoloader": true,
@@ -42,12 +46,10 @@
             "app/Http/helpers.php"
         ],
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
