@@ -1,15 +1,13 @@
 <?php
 
-use App\News;
-
-use function PHPUnit\Framework\assertFalse;
+use App\Models\News;
 
 it('should report expiration status correctly', function () {
-    $news1 = factory(News::class)->make(['expirationDate' => '2000-01-01']);
-    $news2 = factory(News::class)->make(['expirationDate' => '2099-01-01']);
-    $news3 = factory(News::class)->make(['expirationDate' => null]);
+    $news1 = News::factory()->make(['expirationDate' => '2000-01-01']);
+    $news2 = News::factory()->make(['expirationDate' => '2099-01-01']);
+    $news3 = News::factory()->make(['expirationDate' => null]);
 
-    assertTrue($news1->isExpired);
-    assertFalse($news2->isExpired);
-    assertFalse($news3->isExpired);
+    expect($news1->isExpired)->toBeTrue();
+    expect($news2->isExpired)->toBeFalse();
+    expect($news3->isExpired)->toBeFalse();
 });
