@@ -51,7 +51,16 @@ const routes: Array<RouteConfig> = [
         path: "/users",
         name: "User",
         component: () =>
-            import(/* webpackChunkName: "news" */ "../views/User.vue"),
+            import(/* webpackChunkName: "user" */ "../views/User.vue"),
+        meta: { authorize: [Role.Admin] },
+    },
+    {
+        path: "/activities",
+        name: "Activities",
+        component: () =>
+            import(
+                /* webpackChunkName: "activities" */ "../views/Activities.vue"
+            ),
         meta: { authorize: [Role.Admin] },
     },
     {
@@ -74,7 +83,8 @@ router.beforeEach((to, from, next) => {
     if (authorize) {
         if (!currentUser) {
             // not logged in so redirect to login page with the return url
-            return next({ path: "/login", query: { returnUrl: to.path } });
+            const query = to.path == "/" ? {} : { returnUrl: to.path };
+            return next({ path: "/login", query: query });
         }
 
         // check if route is restricted by role
