@@ -1,6 +1,5 @@
 <?php
 
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -13,7 +12,7 @@ use Illuminate\Support\Facades\Route;
 | contains the "web" middleware group. Now create something great!
 |
 */
-Route::get('/startseite', 'PageController@landing');
-Route::get('/files/{name}', 'FilesController@download');
-Auth::routes();
-Route::fallback('PageController@show');
+
+Route::get('/startseite', 'App\Http\Controllers\PageController@landing');
+Route::get('/files/{name}', 'App\Http\Controllers\FilesController@download');
+Route::fallback('App\Http\Controllers\PageController@show');
